requirements.txt
# Core Dependencies (Computer Vision & Image Processing)
opencv-python==4.10.0.84
face-recognition==1.3.0
numpy==1.26.4
Pillow==10.0.1

# Firebase (Cloud Database & Storage)
firebase-admin==6.2.0

# Hardware Control (Raspberry Pi + Fingerprint Scanner)
adafruit-circuitpython-fingerprint
pyserial==3.5

# Database (Optional - if using MySQL)
mysql-connector-python==8.1.0

# Utilities
python-dotenv==1.0.1
